:contruction: add api query string

+++ src/generateApi/index.ts
@@ -1,11 +1,13 @@
 import App from "@/models/App"
 import Model from "@/models/Model"
-import { ModelTypes } from "@/types"
+import { ApiNames, IModel, ModelTypes } from "@/types"
 import express from "express"
 import Joi from "joi"
+import aqp from "api-query-params"
 
 const router = express.Router()
 import { MongoClient, ObjectId } from "mongodb"
+import ApiSchema from "@/models/ApiSchema"
 router.use(express.json())
 
 const url = "mongodb://localhost:27017"
@@ -36,8 +38,48 @@ const mapToJoi = (type: ModelTypes, required: boolean) => {
 	return joi
 }
 
+interface ValidateApiMetodProps {
+	model: IModel
+	params: Record<string, any>
+	methodName: ApiNames
+}
+
+const validateApiMethod = async ({
+	model,
+	params,
+	methodName,
+}: ValidateApiMetodProps) => {
+	const apiSchema = await ApiSchema.findOne({
+		model: model._id,
+	}).lean()
+
+	if (!apiSchema) {
+		return {
+			status: false,
+			code: 404,
+			message: `${params.modelName} not found`,
+		}
+	}
+
+	const method = apiSchema.methods.find((e) => e.name === methodName)!
+
+	if (!method.active) {
+		return {
+			status: false,
+			code: 403,
+			message: "",
+		}
+	}
+
+	return {
+		status: true,
+	}
+}
+
 router.get("/rest/:slug/:modelName", async (req, res) => {
 	const { params } = req
+	const { filter, skip, limit, sort } = aqp(req.query)
+
 	const app = await App.findOne({
 		slug: params.slug,
 	})
@@ -45,6 +87,7 @@ router.get("/rest/:slug/:modelName", async (req, res) => {
 		res.status(404).json({ message: `${params.slug} not found` })
 		return
 	}
+
 	const model = await Model.findOne({
 		$and: [
 			{
@@ -61,13 +104,34 @@ router.get("/rest/:slug/:modelName", async (req, res) => {
 		return
 	}
 
+	const result = await validateApiMethod({
+		model,
+		params,
+		methodName: "GET_ALL",
+	})
+
+	if (!result.status) {
+		res.status(result!.code!).send(result.message)
+		return
+	}
+
 	const dbName = `tront:${app.user}`
 	const colName = `${app._id}:${model._id}`
 	const col = await apiModel({ dbName, colName })
-	const data = await col.find({}).toArray()
+	const data = await col
+		.find()
+		.filter(filter)
+		.skip(skip || 0)
+		.limit(limit || 10)
+		.sort(sort as any)
+		.toArray()
 
 	res.send({
 		data,
+		filter,
+		skip,
+		limit,
+		sort,
 	})
 })
 
@@ -98,6 +162,17 @@ router.get("/rest/:slug/:modelName/:id", async (req, res) => {
 			return
 		}
 
+		const result = await validateApiMethod({
+			model,
+			params,
+			methodName: "GET_ONE",
+		})
+
+		if (!result.status) {
+			res.status(result!.code!).send(result.message)
+			return
+		}
+
 		const dbName = `tront:${app.user}`
 		const colName = `${app._id}:${model._id}`
 		const col = await apiModel({ dbName, colName })
@@ -106,7 +181,12 @@ router.get("/rest/:slug/:modelName/:id", async (req, res) => {
 			_id: ObjectId.createFromHexString(params.id),
 		})
 
-		res.send({ data })
+		if (!data) {
+			res.sendStatus(404)
+			return
+		}
+
+		res.send(data)
 	} catch (error: any) {
 		res.send({ error: error.message })
 	}
@@ -138,6 +218,17 @@ router.post("/rest/:slug/:modelName", async (req, res) => {
 			return
 		}
 
+		const result = await validateApiMethod({
+			model,
+			params,
+			methodName: "POST",
+		})
+
+		if (!result.status) {
+			res.status(result!.code!).send(result.message)
+			return
+		}
+
 		const obj: any = {}
 
 		for (const field of model.fields) {
@@ -160,7 +251,7 @@ router.post("/rest/:slug/:modelName", async (req, res) => {
 		const col = await apiModel({ dbName, colName })
 		const data = await col.insertOne(body)
 
-		res.send({ data })
+		res.status(201).send(await col.findOne({ _id: data.insertedId }))
 	} catch (error: any) {
 		res.status(500).json({
 			error: error.message,
@@ -195,6 +286,17 @@ router.patch("/rest/:slug/:modelName/:id", async (req, res) => {
 			return
 		}
 
+		const result = await validateApiMethod({
+			model,
+			params,
+			methodName: "PATCH",
+		})
+
+		if (!result.status) {
+			res.status(result!.code!).send(result.message)
+			return
+		}
+
 		const obj: any = {}
 
 		for (const field of model.fields) {
@@ -214,12 +316,14 @@ router.patch("/rest/:slug/:modelName/:id", async (req, res) => {
 		const dbName = `tront:${app.user}`
 		const colName = `${app._id}:${model._id}`
 		const col = await apiModel({ dbName, colName })
-		const data = await col.updateOne(
+		await col.updateOne(
 			{ _id: ObjectId.createFromHexString(params.id) },
 			{ $set: body }
 		)
 
-		res.send({ data })
+		res.send(
+			await col.findOne({ _id: ObjectId.createFromHexString(params.id) })
+		)
 	} catch (error: any) {
 		console.log(error.message)
 		res.status(500).json({
@@ -255,15 +359,28 @@ router.delete("/rest/:slug/:modelName/:id", async (req, res) => {
 			return
 		}
 
+		const result = await validateApiMethod({
+			model,
+			params,
+			methodName: "DELETE",
+		})
+
+		if (!result.status) {
+			res.status(result!.code!).send(result.message)
+			return
+		}
+
 		const dbName = `tront:${app.user}`
 		const colName = `${app._id}:${model._id}`
 		const col = await apiModel({ dbName, colName })
 
-		const result = await col.deleteOne({
+		const data = await col.deleteOne({
 			_id: ObjectId.createFromHexString(params.id),
 		})
 
-		res.send({ result })
+		if (!data.deletedCount) return res.status(404)
+
+		res.sendStatus(204)
 	} catch (error) {
 		res.send({ error })
 	}
